backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langgraph>=1.1.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pypdf>=4.0.0
gspread>=6.0.0
google-auth>=2.0.0